backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
PyJWT
python-multipart
httpx
python-dotenv
weaviate-client
sentence-transformers
numpy 